% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefugeesMigrants.R
\docType{data}
\name{RefugeesMigrants}
\alias{RefugeesMigrants}
\title{"RefugeesMigrants"}
\format{
3932 total rows of data and 11 columns
\describe{
\item{date_found_dead}{Data death discovered, in y-m-d format}
\item{number}{Number found dead in the incident in that row}
\item{region_of_origin}{Region of origin of the migrant/refugee(s)}
\item{country_of_incident}{Country where death occured/bodies were found}
\item{inc_men}{Whether the list of the dead included men}
\item{inc_women}{Whether the list of the dead included women}
\item{inc_children}{Whether the list of the dead included children or teenagers}
\item{cause_of_death}{Cause of death from the article}
\item{source}{Data source for the row, usually news}
\item{name_gender_age}{String of name, gender, and age of all the dead in the row, if known}
\item{found_dead}{Original characters for dates, most but not all in y-m-d format}
}
}
\source{
\url{https://github.com/ondata/the-list}
}
\usage{
RefugeesMigrants
}
\description{
A data set of migrants and refugees who died trying to reach Europe
Collected by unitedagainstracism.org and The Guardian
Turned into csv format by Andrea Borusso
as of June, 2018
}
\keyword{datasets}
